backend/stt_wispher/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==1.10.11
openai_whisper==20230314
numpy==1.25
slowapi
pymongo
PyYAML==6.0.1
logger
cachetools